Added checksum, catch exception

+++ dspace-api/src/main/java/org/dspace/storage/bitstore/AzureBitStoreService.java
@@ -22,6 +22,9 @@ import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.security.DigestInputStream;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.util.Map;
 //import com.azure.storage.common.StorageSharedKeyCredential;
 
@@ -91,23 +94,37 @@ public class AzureBitStoreService implements BitStoreService{
 
     @Override
     public void put(Bitstream bitstream, InputStream in) throws IOException {
-        String key = getFullKey(bitstream.getInternalId());
         //Copy istream to temp file, and send the file, with some metadata
-        //File scratchFile = File.createTempFile(bitstream.getInternalId(), "blobbs");
         File scratchFile = getTempFile(bitstream);
-        String fileName = scratchFile.getName();
-        FileUtils.copyInputStreamToFile(in, scratchFile);
-        long contentLength = scratchFile.length();
-        BlobContainerClient containerClient;
-        containerClient = blobServiceClient.getBlobContainerClient(containerName);
-        BlobClient blobClient = containerClient.getBlobClient(fileName);
-        blobClient.uploadFromFile(String.valueOf(scratchFile));
-        scratchFile.getName();
-        bitstream.setSizeBytes(contentLength);
-        //bitstream.setChecksum(putObjectResult.getETag());
-        bitstream.setChecksumAlgorithm(CSA);
-        scratchFile.delete();
-        log.info("Azure mock put");
+        try {
+
+            FileUtils.copyInputStreamToFile(in, scratchFile);
+
+            long contentLength = scratchFile.length();
+            BlobContainerClient containerClient;
+            containerClient = blobServiceClient.getBlobContainerClient(containerName);
+
+            String fileName = scratchFile.getName();
+            BlobClient blobClient = containerClient.getBlobClient(fileName);
+            blobClient.uploadFromFile(String.valueOf(scratchFile));
+            bitstream.setSizeBytes(contentLength);
+            try {
+                DigestInputStream dis = new DigestInputStream(in, MessageDigest.getInstance(CSA));
+                bitstream.setChecksum(Utils.toHex(dis.getMessageDigest().digest()));
+                bitstream.setChecksumAlgorithm(CSA);
+                scratchFile.delete();
+            } catch (NoSuchAlgorithmException e) {
+                // Should never happen
+                log.warn("Caught NoSuchAlgorithmException", e);
+            }
+        } catch (Exception e) {
+            log.error("put(" + bitstream.getInternalId() + ", inputstream)", e);
+            throw new IOException(e);
+        }finally {
+            if (scratchFile.exists()) {
+                scratchFile.delete();
+            }
+        }
     }
 
     @Override
